Test of service impl component

+++ src/main/java/com/company/Calulator/Service/ResultService.java
@@ -5,6 +5,6 @@ import com.company.Calulator.DTO.ResultObject;
 import java.math.BigDecimal;
 
 public interface ResultService {
-    ResultObject sum (BigDecimal operator_1, BigDecimal operator_2);
-    ResultObject subtract (BigDecimal operator_1, BigDecimal operator_2);
+    ResultObject sum (Float operator_1, Float operator_2);
+    ResultObject subtract (Float operator_1, Float operator_2);
 }
